% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_tau_interval}
\alias{check_tau_interval}
\title{Positive whole number tester}
\usage{
check_tau_interval(input, descriptor)
}
\arguments{
\item{input}{Candidate percentage}
}
\value{
Fail if not within 0 and 1
}
\description{
Positive whole number tester
}
\keyword{internal}
