% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_user_input}
\alias{check_user_input}
\title{User input checker}
\usage{
check_user_input(
  data,
  grouping_variable,
  sum_score,
  minimum_item_response,
  maximum_item_response,
  number_items
)
}
\arguments{
\item{data}{A dataset, ideally a data.frame.}

\item{grouping_variable}{The name of the grouping_variable in the dataset.}

\item{sum_score}{The name of the sum_score variable in the dataset. This
variable must be whole numbers.}

\item{minimum_item_response}{Theorical minimum value on items summed to
create sum score. For example, with Likert data with 5 response categories
coded 0,1,2,3,4, this argument would be \code{minimum_item_response = 0}.}

\item{maximum_item_response}{Theorical maximum value on items summed to
create sum score. For example, with Likert data with 3 response categories
coded 1,2,3, this argument would be \code{maximum_item_response = 3}.}

\item{number_items}{Number of items summed to create the sum score.}
}
\value{
Check user input
}
\description{
User input checker
}
\keyword{internal}
