% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_splines}
\alias{create_splines}
\title{Create Spline object}
\usage{
create_splines(user_input)
}
\arguments{
\item{user_input}{additional user input}
}
\value{
Return spline
}
\description{
Create Spline object
}
\keyword{internal}
