% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Create Stan data list object}
\usage{
create_stan_data(y_ord, grp, user_input)
}
\arguments{
\item{y_ord}{numeric input}

\item{grp}{grouping variable}

\item{user_input}{additional user input}
}
\value{
Data list object for Stan
}
\description{
Create Stan data list object
}
\keyword{internal}
