% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generic_data_checks}
\alias{generic_data_checks}
\title{Generic data checks}
\usage{
generic_data_checks(data, grouping_variable, sum_score)
}
\arguments{
\item{data}{A dataset, ideally a data.frame.}

\item{grouping_variable}{The name of the grouping_variable in the dataset.}

\item{sum_score}{The name of the sum_score variable in the dataset. This
variable must be whole numbers.}
}
\value{
Data
}
\description{
Generic data checks
}
\keyword{internal}
