% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generic_numeric_checks}
\alias{generic_numeric_checks}
\title{Generic numeric checks}
\usage{
generic_numeric_checks(
  dt,
  minimum_item_response,
  maximum_item_response,
  number_items
)
}
\arguments{
\item{dt}{Modified form of original data}

\item{minimum_item_response}{Theorical minimum value on items summed to
create sum score. For example, with Likert data with 5 response categories
coded 0,1,2,3,4, this argument would be \code{minimum_item_response = 0}.}

\item{maximum_item_response}{Theorical maximum value on items summed to
create sum score. For example, with Likert data with 3 response categories
coded 1,2,3, this argument would be \code{maximum_item_response = 3}.}

\item{number_items}{Number of items summed to create the sum score.}
}
\value{
The adjusted sum scores
}
\description{
Generic numeric checks
}
\keyword{internal}
