% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_positive_whole_number}
\alias{is_positive_whole_number}
\title{Positive whole number tester}
\usage{
is_positive_whole_number(input)
}
\arguments{
\item{input}{Candidate whole number}
}
\value{
Fail if not non-negative whole number
}
\description{
Positive whole number tester
}
\keyword{internal}
