% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_whole_number}
\alias{is_whole_number}
\title{Whole number tester}
\usage{
is_whole_number(input)
}
\arguments{
\item{input}{Candidate whole number}
}
\value{
Fail if not whole number
}
\description{
Whole number tester
}
\keyword{internal}
