% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugcflss_compute_stats.R
\name{ugcflss_compute_ps}
\alias{ugcflss_compute_ps}
\title{Compute Probability of Superiority}
\usage{
ugcflss_compute_ps(res_obj, which_groups)
}
\arguments{
\item{res_obj}{Object returned by main function}

\item{which_groups}{Two group ids}
}
\value{
Returns PS.
}
\description{
Compute Probability of Superiority
}
\keyword{internal}
