% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugcflss_compute_stats.R
\name{ugcflss_compute_stats}
\alias{ugcflss_compute_stats}
\title{Compute some requested statistic}
\usage{
ugcflss_compute_stats(res_obj, which_groups, stat = "pmf", tau = 0.5)
}
\arguments{
\item{res_obj}{Object returned by main function}

\item{which_groups}{One or more group id(s) (integer)}

\item{stat}{One of "pmf", "ccdf", "quantile", "mean", or "sd"}

\item{tau}{The probability for quantile calculations}
}
\value{
Returns requested statistic.
}
\description{
Compute some requested statistic
}
\keyword{internal}
