% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugcflss_simple_desc.R
\name{ugcflss_describe_plot}
\alias{ugcflss_describe_plot}
\title{Plot requested statistic for entire sample and by group}
\usage{
ugcflss_describe_plot(res_obj, stat = "median", interval = 0.89, tau = 0.5)
}
\arguments{
\item{res_obj}{Object returned by main function}

\item{stat}{Statistic of interest, one of "median", "mean", "sd"
or "quantile".}

\item{interval}{Some quantile interval between 0 and 1}

\item{tau}{If stat = "quantile", the value of the probability between 0 and 1}
}
\value{
Returns plot
}
\description{
Plot requested statistic for entire sample and by group
}
