% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugcflss_ccdf.R
\name{ugcflss_exceed_all_plot}
\alias{ugcflss_exceed_all_plot}
\title{Plot CCDF for entire sample}
\usage{
ugcflss_exceed_all_plot(res_obj, interval = 0.89, show_intervals = FALSE)
}
\arguments{
\item{res_obj}{Object returned by main function}

\item{interval}{Some quantile interval of interest between 0 and 1}

\item{show_intervals}{If TRUE, show intervals on plot,
if FALSE, do not show intervals}
}
\value{
Returns plot
}
\description{
Plot CCDF for entire sample
}
