% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugcflss_ccdf.R
\name{ugcflss_exceed_group}
\alias{ugcflss_exceed_group}
\title{Compute CCDF by group}
\usage{
ugcflss_exceed_group(res_obj, interval = 0.89, convergence = FALSE)
}
\arguments{
\item{res_obj}{Object returned by main function}

\item{interval}{Some quantile interval of interest between 0 and 1}

\item{convergence}{If TRUE, returns CCDF convergence statistics.
If FALSE, creates dataset showing CCDF which can be used for plotting
the CCDF.}
}
\value{
Returns dataset.
}
\description{
Compute CCDF by group
}
