% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugcflss_fit_model.R
\name{ugcflss_fit_model}
\alias{ugcflss_fit_model}
\title{Analyze a dataset}
\usage{
ugcflss_fit_model(
  data,
  grouping_variable = NA_character_,
  sum_score = NA_character_,
  minimum_item_response = NA_integer_,
  maximum_item_response = NA_integer_,
  number_items = NA_integer_,
  warmup = 750,
  sampling = 750,
  refresh = max((warmup + sampling)\%/\%10, 1),
  adapt_delta = 0.9,
  max_treedepth = 10,
  chains = 3,
  cores = min(chains, max(parallel::detectCores() - 2, 1)),
  seed = sample.int(.Machine$integer.max, 1),
  override_twenty_groups = FALSE,
  show_messages = TRUE
)
}
\arguments{
\item{data}{A dataset, ideally a data.frame.}

\item{grouping_variable}{The name of the grouping_variable in the dataset.}

\item{sum_score}{The name of the sum_score variable in the dataset. This
variable must be whole numbers.}

\item{minimum_item_response}{Theorical minimum value on items summed to
create sum score. For example, with Likert data with 5 response categories
coded 0,1,2,3,4, this argument would be \code{minimum_item_response = 0}.}

\item{maximum_item_response}{Theorical maximum value on items summed to
create sum score. For example, with Likert data with 3 response categories
coded 1,2,3, this argument would be \code{maximum_item_response = 3}.}

\item{number_items}{Number of items summed to create the sum score.
Must be at least 1.}

\item{warmup}{Number of iterations used to warmup the sampler, per chain.}

\item{sampling}{Number of iterations retained for inference, per chain.}

\item{refresh}{(Positive whole number) How often to print the status of
the sampler.}

\item{adapt_delta}{Number in (0,1). Increase to resolve divergent
transitions.}

\item{max_treedepth}{(Positive whole number) Increase to resolve problems
with maximum tree depth.}

\item{chains}{Number of chains to use.}

\item{cores}{Number of cores to use.}

\item{seed}{Random seed.}

\item{override_twenty_groups}{By default, we assume your grouping_variable
does not have more than 20 levels. If you want to override this default,
add \code{override_twenty_groups = TRUE} to your function call.}

\item{show_messages}{(Logical) If TRUE, show messages from Stan sampler,
if FALSE, hide messages.}
}
\value{
Objects containing analysis results.
}
\description{
Analyze a dataset
}
\examples{
\dontrun{
model_1 <- ugcflss_fit_model(
  data, grouping_variable, sum_score,
  minimum_item_response, maximum_item_response, number_items
)
}
}
