% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugcflss_simple_desc.R
\name{ugcflss_pairwise_plot}
\alias{ugcflss_pairwise_plot}
\title{Perform pairwise comparisons of requested statistic}
\usage{
ugcflss_pairwise_plot(
  res_obj,
  stat = "median",
  interval = 0.89,
  tau = 0.5,
  comparison = "difference"
)
}
\arguments{
\item{res_obj}{Object returned by main function}

\item{stat}{Statistic of interest, one of "median", "mean", "sd"
or "quantile".}

\item{interval}{Some quantile interval between 0 and 1}

\item{tau}{If stat = "quantile", the value of the probability between 0 and 1}

\item{comparison}{One of difference, ratio or log-ratio. Log-ratio
will lead to an error if any of the values to be logged are negative
though this should not happen.}
}
\value{
Returns dataset
}
\description{
Perform pairwise comparisons of requested statistic
}
