% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugcflss_ppc.R
\name{ugcflss_ppd}
\alias{ugcflss_ppd}
\title{Compute posterior predictive distribution}
\usage{
ugcflss_ppd(res_obj, ppd_samples = 200)
}
\arguments{
\item{res_obj}{Object returned by main function}

\item{ppd_samples}{Number of iterations to use to obtain
the posterior predictive distribution.}
}
\value{
Returns dataset
}
\description{
Compute posterior predictive distribution
}
