% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugcflss_ppc.R
\name{ugcflss_ppd_plot}
\alias{ugcflss_ppd_plot}
\title{Basic PPD density overlay plot}
\usage{
ugcflss_ppd_plot(res_obj, ppd_samples = 200, by_group = FALSE)
}
\arguments{
\item{res_obj}{Object returned by main function}

\item{ppd_samples}{Number of iterations to use to obtain
the posterior predictive distribution.}

\item{by_group}{If TRUE, plot posterior predictive distribution by group.
If FALSE, plot for entire sample.}
}
\value{
Returns density overlay plot
}
\description{
Basic PPD density overlay plot
}
