% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ugcflss-package.R
\name{ugcflss-package}
\alias{ugcflss-package}
\alias{ugcflss}
\title{The 'ugcflss' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
